improve naming and use SqsDeletionConsumer helper object
